sources

+++ codebase/android/app/src/main/java/com/makeappssimple/abhimanyu/financemanager/android/ui/screens/add_source/AddSourceViewModel.kt
@@ -42,7 +42,7 @@ class AddSourceViewModel @Inject constructor(
             sourceRepository.insertSource(
                 source = Source(
                     balanceAmount = Amount(
-                        value = balanceAmount.toFloat(),
+                        value = balanceAmount.toLong(),
                     ),
                     type = type,
                     name = name,
